src/requirements.txt
Flask>=3.0,<4.0
python-plexapi>=4.15,<5.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
waitress>=3.0,<4.0
